UI Move to common folder
