+++ sources/packages/web/src/components/common/students/applicationTracker/Completed.vue
@@ -220,10 +220,10 @@ import {
 import { onMounted, ref, defineComponent, computed } from "vue";
 import { ApplicationService } from "@/services/ApplicationService";
 import { CompletedApplicationDetailsAPIOutDTO } from "@/services/http/dto/Application.dto";
-import ApplicationStatusTrackerBanner from "@/components/students/applicationTracker/generic/ApplicationStatusTrackerBanner.vue";
-import DisbursementBanner from "@/components/students/applicationTracker/DisbursementBanner.vue";
-import MultipleDisbursementBanner from "@/components/students/applicationTracker/MultipleDisbursementBanner.vue";
-import RelatedApplicationChanged from "@/components/students/applicationTracker/RelatedApplicationChanged.vue";
+import ApplicationStatusTrackerBanner from "@/components/common/students/applicationTracker/generic/ApplicationStatusTrackerBanner.vue";
+import DisbursementBanner from "@/components/common/students/applicationTracker/DisbursementBanner.vue";
+import MultipleDisbursementBanner from "@/components/common/students/applicationTracker/MultipleDisbursementBanner.vue";
+import RelatedApplicationChanged from "@/components/common/students/applicationTracker/RelatedApplicationChanged.vue";
 import { StudentRoutesConst } from "@/constants/routes/RouteConstants";
 import { useRouter } from "vue-router";
 import {
