+++ sources/packages/web/src/components/common/students/applicationTracker/MultipleDisbursementBanner.vue
@@ -128,7 +128,7 @@
   />
 </template>
 <script lang="ts">
-import ApplicationStatusTrackerBanner from "@/components/students/applicationTracker/generic/ApplicationStatusTrackerBanner.vue";
+import ApplicationStatusTrackerBanner from "@/components/common/students/applicationTracker/generic/ApplicationStatusTrackerBanner.vue";
 import { COEStatus, DisbursementScheduleStatus } from "@/types";
 import { defineComponent, PropType, computed } from "vue";
