+++ sources/packages/web/src/components/common/students/applicationTracker/Submitted.vue
@@ -15,7 +15,7 @@
 </template>
 <script lang="ts">
 import { defineComponent } from "vue";
-import ApplicationStatusTrackerBanner from "@/components/students/applicationTracker/generic/ApplicationStatusTrackerBanner.vue";
+import ApplicationStatusTrackerBanner from "@/components/common/students/applicationTracker/generic/ApplicationStatusTrackerBanner.vue";
 export default defineComponent({
   components: {
     ApplicationStatusTrackerBanner,
