+++ sources/packages/web/src/components/common/students/applicationTracker/generic/ApplicationStatusTrackerBanner.vue
